collections defaultdict 1.4

+++ Intermediate/Collections.py
@@ -1,5 +1,5 @@
 from collections import defaultdict
-d = defaultdict(int)
+d = defaultdict(float)
 d["a"] = 1
 d["b"] = 3
 print(d["d"])
